Use pinned Node version in dockerfiles
Ideally we should keep using the `14` tag because it would mean that we
are always using the latest release of that major, but currently this
isn't true for cypress' images. Right now, using the `14` tag installs
Node at version 14.7.0 which is incompatible with a few of our
dependencies, and the immediate fix is to use the latest minor available
to this specific image (which is 14.19.0 – even though Node already has
a release newer than it, latest for v14 is 14.19.1).